+++ pipeline-analyze.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-echo "Analyze code in repository"
-./gradlew dependencyCheckAggregate || exit
-./gradlew jacocoTestReport || exit
-./gradlew sonar || exit