+++ pipeline-clean-repo.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-echo "Clean the repository"
-./gradlew clean || exit