+++ pipeline-run-system-tests.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-echo "Running System Tests"
-./gradlew :system-tests:check -Psystem-tests --stacktrace || exit